Structure files as python package
