+++ .gitignore
@@ -1,0 +1,5 @@
+/build
+/*.egg-info
+
+*.pyc
+
