+++ siyi_sdk/cameras.py
@@ -1,0 +1,17 @@
+"""
+This script defines the camera specs
+"""
+
+class A8MINI:
+    MAX_YAW_DEG = 135.0
+    MIN_YAW_DEG = -135.0
+    MAX_PITCH_DEG = 25.0
+    MIN_PITCH_DEG = -90.0
+    MAX_ZOOM = 6.0
+
+class ZR10:
+    MAX_YAW_DEG = 135.0
+    MIN_YAW_DEG = -135.0
+    MAX_PITCH_DEG = 25.0
+    MIN_PITCH_DEG = -90.0
+    MAX_ZOOM = 30.0 # 10 optical * 3 digital
